requirements.txt
alembic==1.7.3
astroid==2.8.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
ecdsa==0.17.0
flake8==3.9.2
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
pep8==1.7.1
platformdirs==2.3.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.11.1
python-dotenv==0.19.0
python-jose==3.3.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
